[Feat]: BidService

+++ src/main/java/com/backend/domain/bid/dto/BidResponseDto.java
@@ -3,9 +3,9 @@ package com.backend.domain.bid.dto;
 import java.time.LocalDateTime;
 
 public record BidResponseDto (
-  Long id,
-  Long productId,
-  Long bidderId,
+  Integer id,
+  Integer productId,
+  Integer bidderId,
   long price,
   String status,
   LocalDateTime createDate
